Refactor pre-push synchronization logic and introduce modular components
This commit refactors the pre-push synchronization logic by breaking it down into modular components, enhancing maintainability and readability. New files are added for handling commit selection, current state replication, history replication, and SQL helper functions. Constants and types related to pre-push operations are also defined in dedicated files. The main synchronization function is updated to utilize these new components, improving the overall structure of the code. Tests are included to ensure the correctness of the new implementations and to validate the parsing of pre-push updates.
