+++ bitloops/src/host/checkpoints/strategy/manual_commit/strategy_impl/pre_push_sync/types.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(super) struct PrePushRefUpdate {
+    pub(super) local_ref: String,
+    pub(super) local_sha: String,
+    pub(super) remote_ref: String,
+    pub(super) remote_sha: String,
+    pub(super) local_branch: Option<String>,
+    pub(super) remote_branch: String,
+}
